@@ -1,6 +1,6 @@
 module github.com/nbvghost/dandelion
 
-go 1.17
+go 1.18
 
 require (
 	github.com/PuerkitoBio/goquery v1.8.0
