@@ -9,6 +9,7 @@ import (
 	"github.com/nbvghost/dandelion/entity/extends"
 	"github.com/nbvghost/dandelion/library/util"
 	"github.com/pkg/errors"
+	"strings"
 
 	"net/http"
 	"reflect"
@@ -32,11 +33,13 @@ func (m *Info) GetWithoutAuth() bool {
 
 type emptyViewBase struct {
 	extends.ViewBase
+	Data any
 }
 
-func NewViewResult(name string) constrain.IViewResult {
+func NewViewResult(name string, data any) constrain.IViewResult {
 	return &emptyViewBase{
 		ViewBase: extends.ViewBase{Name: name},
+		Data:     data,
 	}
 }
 
@@ -77,39 +80,37 @@ func (m *service) CheckRoute(isApi bool, route string) (*Info, bool) {
 	var ok bool
 	if isApi {
 		if routeInfo, ok = m.Routes[route]; !ok {
-			return nil, false
+			if routeInfo, ok = m.Routes["/api/"]; !ok {
+				return nil, false
+			}
 		}
 	} else {
 		if routeInfo, ok = m.ViewRoutes[route]; !ok {
-			return nil, false
+			if routeInfo, ok = m.ViewRoutes[""]; !ok {
+				return nil, false
+			}
 		}
 	}
 	return routeInfo, ok
 }
-func (m *service) Handle(context constrain.IContext, isApi bool, route string, binddataFunc func(apiHandler interface{}) error) (bool, bool, interface{}, error) {
-	var hasRoute bool
-	routeInfo, ok := m.CheckRoute(isApi, route)
+
+func (m *service) CreateHandle(isApi bool, pathTemplate string) (any, bool, error) {
+	routeInfo, ok := m.CheckRoute(isApi, pathTemplate)
 	if !ok {
-		return true, hasRoute, nil, action.NewCodeWithError(action.NotFoundRoute, errors.Errorf("没有找到路由:%s", route))
+		return nil, false, errors.Errorf("没有找到路由:%s", pathTemplate)
 	}
-	hasRoute = true
-
 	apiHandler := reflect.New(routeInfo.GetHandlerType()).Interface()
-	if binddataFunc != nil {
-		err := binddataFunc(apiHandler)
-		if err != nil {
-			return true, hasRoute, apiHandler, err
-		}
-	}
-
+	return apiHandler, routeInfo.GetWithoutAuth(), nil
+}
+func (m *service) Handle(context constrain.IContext, withoutAuth bool, routeHandler any) (bool, error) {
 	if m.mappingCallback != nil {
-		if err := m.mappingCallback.Before(context, apiHandler); err != nil {
-			return true, hasRoute, nil, err
+		if err := m.mappingCallback.Before(context, routeHandler); err != nil {
+			return true, err
 		}
 	}
 
-	if routeInfo.WithoutAuth {
-		return false, hasRoute, apiHandler, nil
+	if withoutAuth {
+		return false, nil
 	}
 
 	//interceptors 是有状态的，不支持mapping
@@ -121,68 +122,77 @@ func (m *service) Handle(context constrain.IContext, isApi bool, route string, b
 				for i := range m.interceptors[k] {
 					if m.mappingCallback != nil {
 						if err := m.mappingCallback.Before(context, m.interceptors[k][i]); err != nil {
-							return true, hasRoute, nil, err
+							return true, err
 						}
 					}
 					broken, err := m.interceptors[k][i].Execute(context)
 					if err != nil {
-						return true, hasRoute, nil, err
+						return true, err
 					}
 					if broken {
-						return true, hasRoute, nil, nil
+						return true, nil
 					}
 				}
 			}
 		}
 	}
 
-	if !routeInfo.GetWithoutAuth() {
+	if !withoutAuth {
 		if context.UID() == 0 {
-			return true, hasRoute, nil, action.NewCodeWithError(action.AuthError, errors.New("用户没有授权"))
+			return true, action.NewCodeWithError(action.AuthError, errors.New("用户没有授权"))
 		}
 	}
 
-	return false, hasRoute, apiHandler, nil
+	return false, nil
 
 }
 
-func (m *service) RegisterRoute(path string, handler constrain.IHandler, withoutAuth ...bool) {
-	path = "/" + path
-	if _, ok := m.Routes[path]; ok {
-		panic(errors.New(fmt.Sprintf("存在相同的路由:%s", path)))
+func (m *service) RegisterRoute(pathTemplate string, handler constrain.IHandler, withoutAuth ...bool) {
+	if strings.EqualFold(pathTemplate, "*") {
+		pathTemplate = "/api/"
+	} else {
+		pathTemplate = "/api/" + pathTemplate
+	}
+	if _, ok := m.Routes[pathTemplate]; ok {
+		panic(errors.New(fmt.Sprintf("存在相同的路由:%s", pathTemplate)))
 	}
 	var _withoutAuth bool
 	if len(withoutAuth) > 0 {
 		_withoutAuth = withoutAuth[0]
 	}
-	m.router.HandleFunc("/api"+path, func(writer http.ResponseWriter, request *http.Request) {
+	m.router.HandleFunc(pathTemplate, func(writer http.ResponseWriter, request *http.Request) {
 
 	})
-	m.Routes[path] = &Info{
+	m.Routes[pathTemplate] = &Info{
 		HandlerType: reflect.TypeOf(handler).Elem(),
 		WithoutAuth: _withoutAuth,
 	}
 }
 
-func (m *service) RegisterView(path string, handler constrain.IViewHandler, result constrain.IViewResult, withoutAuth ...bool) {
-	path = "/" + path
-	if _, ok := m.ViewRoutes[path]; ok {
-		panic(errors.New(fmt.Sprintf("存在相同的路由:%s", path)))
+// RegisterView path 为 * 号时，匹配所有没有定义的路由
+func (m *service) RegisterView(pathTemplate string, handler constrain.IViewHandler, result constrain.IViewResult, withoutAuth ...bool) {
+	if strings.EqualFold(pathTemplate, "*") {
+		pathTemplate = ""
+	} else {
+		pathTemplate = "/" + pathTemplate
+	}
+	if _, ok := m.ViewRoutes[pathTemplate]; ok {
+		panic(errors.New(fmt.Sprintf("存在相同的路由:%s", pathTemplate)))
 	}
 	var _withoutAuth bool
 	if len(withoutAuth) > 0 {
 		_withoutAuth = withoutAuth[0]
 	}
-	m.router.HandleFunc(path, func(writer http.ResponseWriter, request *http.Request) {
+	m.router.HandleFunc(pathTemplate, func(writer http.ResponseWriter, request *http.Request) {
 
 	})
-	m.ViewRoutes[path] = &Info{
+	m.ViewRoutes[pathTemplate] = &Info{
 		HandlerType: reflect.TypeOf(handler).Elem(),
 		WithoutAuth: _withoutAuth,
 	}
 	gobext.Register(result)
 }
 
-func New(router *mux.Router, redis constrain.IRedis, mappingCallback constrain.IMappingCallback) IRoute {
+func New(router *mux.Router, redis constrain.IRedis, mappingCallback constrain.IMappingCallback) constrain.IRoute {
 	return &service{router: router, Routes: map[string]*Info{}, ViewRoutes: map[string]*Info{}, redis: redis, mappingCallback: mappingCallback, interceptors: make(map[string][]constrain.IInterceptor)}
 }
