@@ -13,10 +13,22 @@ func (service ContentService) ListContentItemByOID(OID types.PrimaryKey) []model
 	Orm.Model(model.ContentItem{}).Where(map[string]interface{}{"OID": OID}).Order(`"Sort"`).Order(`"UpdatedAt" desc`).Find(&menus)
 	return menus
 }
-func (service ContentService) GetContentItemDefault(db *gorm.DB, OID types.PrimaryKey) *model.ContentItem {
+func (service ContentService) GetContentItemByTypeTemplateName(db *gorm.DB, OID types.PrimaryKey, typ model.ContentTypeType, templateName string) *model.ContentItem {
 	Orm := db
 	var contentItem model.ContentItem
-	Orm.Model(&model.ContentItem{}).Where("OID=? And Type=?", OID, model.ContentTypeProducts).First(&contentItem)
+	Orm.Model(&model.ContentItem{}).Where(`"OID"=? And "Type"=? And "TemplateName"=?`, OID, typ, templateName).First(&contentItem)
+	return &contentItem
+}
+func (service ContentService) GetContentItemOfIndex(db *gorm.DB, OID types.PrimaryKey) *model.ContentItem {
+	Orm := db
+	var contentItem model.ContentItem
+	Orm.Model(&model.ContentItem{}).Where(`"OID"=? And "Type"=?`, OID, model.ContentTypeIndex).First(&contentItem)
+	return &contentItem
+}
+func (service ContentService) GetContentItemOfProducts(db *gorm.DB, OID types.PrimaryKey) *model.ContentItem {
+	Orm := db
+	var contentItem model.ContentItem
+	Orm.Model(&model.ContentItem{}).Where(`"OID"=? And "Type"=?`, OID, model.ContentTypeProducts).First(&contentItem)
 	if contentItem.ID == 0 {
 		return nil
 	}
